+++ swing/framework-model/src/main/java/is/codion/swing/framework/model/SwingEntityEditor.java
@@ -21,6 +21,7 @@ package is.codion.swing.framework.model;
 import is.codion.common.utilities.proxy.ProxyBuilder;
 import is.codion.common.utilities.proxy.ProxyBuilder.ProxyMethod;
 import is.codion.framework.db.EntityConnectionProvider;
+import is.codion.framework.domain.entity.EntityDefinition;
 import is.codion.framework.domain.entity.EntityType;
 import is.codion.framework.domain.entity.attribute.Attribute;
 import is.codion.framework.domain.entity.attribute.Column;
@@ -191,7 +192,7 @@ public final class SwingEntityEditor extends DefaultEntityEditor {
 		 * @see ForeignKeyDefinition#attributes()
 		 */
 		public EntityComboBoxModel create(ForeignKey foreignKey) {
-			return componentModels.createComboBoxModel(requireNonNull(foreignKey), editor);
+			return componentModels.createComboBoxModel(requireNonNull(foreignKey), editor.connectionProvider());
 		}
 
 		/**
@@ -202,7 +203,7 @@ public final class SwingEntityEditor extends DefaultEntityEditor {
 		 * @see FilterComboBoxModel#NULL_CAPTION
 		 */
 		public <T> FilterComboBoxModel<T> create(Column<T> column) {
-			return componentModels.createComboBoxModel(requireNonNull(column), editor);
+			return componentModels.createComboBoxModel(requireNonNull(column), editor.connectionProvider());
 		}
 	}
 
@@ -223,7 +224,7 @@ public final class SwingEntityEditor extends DefaultEntityEditor {
 		 * null item caption if the underlying attribute is nullable.
 		 * <p>If the foreign key has select attributes defined, those are set in the combo box model.
 		 * @param foreignKey the foreign key for which to create a {@link EntityComboBoxModel}
-		 * @param editor the editor
+		 * @param connectionProvider the connection provider
 		 * @return a {@link EntityComboBoxModel} for the given foreign key
 		 * @see FilterComboBoxModel#NULL_CAPTION
 		 * @see EntityComboBoxModel.Builder#nullCaption(String)
@@ -232,14 +233,15 @@ public final class SwingEntityEditor extends DefaultEntityEditor {
 		 * @see EntityComboBoxModel.Builder#attributes(Collection)
 		 * @see ForeignKeyDefinition#attributes()
 		 */
-		public EntityComboBoxModel createComboBoxModel(ForeignKey foreignKey, SwingEntityEditor editor) {
-			ForeignKeyDefinition foreignKeyDefinition = requireNonNull(editor).entityDefinition().foreignKeys().definition(foreignKey);
+		public EntityComboBoxModel createComboBoxModel(ForeignKey foreignKey, EntityConnectionProvider connectionProvider) {
+			EntityDefinition entityDefinition = requireNonNull(connectionProvider).entities().definition(requireNonNull(foreignKey).entityType());
+			ForeignKeyDefinition foreignKeyDefinition = entityDefinition.foreignKeys().definition(foreignKey);
 
 			return EntityComboBoxModel.builder()
 							.entityType(foreignKey.referencedType())
-							.connectionProvider(editor.connectionProvider())
+							.connectionProvider(connectionProvider)
 							.attributes(foreignKeyDefinition.attributes())
-							.includeNull(editor.nullable(foreignKey))
+							.includeNull(entityDefinition.foreignKeys().nullable(foreignKey))
 							.build();
 		}
 
@@ -248,16 +250,17 @@ public final class SwingEntityEditor extends DefaultEntityEditor {
 		 * This default implementation returns a sorted {@link FilterComboBoxModel} using the default
 		 * null item caption if the underlying column is nullable
 		 * @param column the column
-		 * @param editor the editor
+		 * @param connectionProvider the connection provider
 		 * @param <T> the value type
 		 * @return a combo box model based on the given column
 		 * @see FilterComboBoxModel#NULL_CAPTION
 		 */
-		public <T> FilterComboBoxModel<T> createComboBoxModel(Column<T> column, SwingEntityEditor editor) {
-			boolean nullable = requireNonNull(editor).nullable(column);
+		public <T> FilterComboBoxModel<T> createComboBoxModel(Column<T> column, EntityConnectionProvider connectionProvider) {
+			EntityDefinition entityDefinition = requireNonNull(connectionProvider).entities().definition(requireNonNull(column).entityType());
+			boolean nullable = entityDefinition.columns().definition(column).nullable();
 
 			return FilterComboBoxModel.builder()
-							.items(new ColumnItems<>(editor.connectionProvider(), column))
+							.items(() -> connectionProvider.connection().select(column))
 							.nullItem(nullable ? createNullItem(column) : null)
 							.includeNull(nullable)
 							.build();